% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{build}
\alias{build}
\title{Build HTML report from template}
\usage{
build(plotter, template)
}
\arguments{
\item{plotter}{An object of class "htmlReport".}

\item{template}{A character string specifying the file path of the template.}
}
\description{
Build HTML report from template
}
