% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{compress_data}
\alias{compress_data}
\title{encode and compress data json}
\arguments{
\item{data}{data to compress}
}
\value{
encoded and compressed json
}
\description{
compress_data
}
\examples{
test_data <- "testdatatestdatatestdata"
plotter <- htmlReport$new()
plotter$compress_data(list(data = test_data))
}
