% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_standard_triplet_df}
\alias{.prepare_standard_triplet_df}
\title{.prepare_standard_triplet_df}
\usage{
.prepare_standard_triplet_df(
  list_content,
  list_levels = NULL,
  list_types = NULL,
  default_type = "ul"
)
}
\arguments{
\item{list_content}{Vector or list of elements with which to build the list}

\item{list_levels}{Vector or list defining nesting levels. Default NULL}

\item{list_types}{Vector or list with types to assign to each element ("ul"
for unordered or "ol" for ordered).' If not supplied, default_type will
determine list type for each element. Default NULL}

\item{default_type}{List type to default if list_types is undefined, "ul"
for unordered or "ol" for ordered. Default "ul"}
}
\value{
Data frame of contents, levels and types, ready for make_html_list.
}
\description{
Build triplet data frame for make_html_list function from three vectors.
User should NEVER have to worry about this function
}
