% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{extract_data}
\alias{extract_data}
\title{Retrieve Data from htmlReport Object}
\arguments{
\item{options}{A list containing options for data retrieval.}
}
\value{
A list containing the retrieved data and additional information.
}
\description{
This method retrieves data from an \code{htmlReport} object
based on specified options.
}
\details{
Retrieve Data from htmlReport Object
}
\examples{
test_df <- data.frame(matrix(1, nrow = 5, ncol = 5))
test_df[1, ] <- letters[1:5]
test_df[, 1] <- letters[1:5]
plotter <- htmlReport$new(container = list(test_df = test_df))
plotter$extract_data(options = list(id = "test_df"))
}
