% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_col_n_row_span-htmlReport-method}
\alias{get_col_n_row_span-htmlReport-method}
\title{converts span noted table into a span map.}
\arguments{
\item{table}{Colspan and rowspan maps.}
}
\value{
A list of two tables: colspans and rowspans.
}
\description{
get_col_n_row_span
}
\details{
What we mean by "span map" is a table with the same dimensions
as the input table, but with a number in each cell specifyin its column
and row spans, respectively. A colspan of three means that it will merge with
the two cells on its right into the same cell in the final table. Same logic
for row span map, it will merge with the two cells below it.
}
\examples{
test_table <- data.frame(c("A", "B", "C"), c("colspan", "E", "F"))
plotter <- htmlReport$new()
plotter$get_col_n_row_span(test_table)
}
