% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_data_for_plot}
\alias{get_data_for_plot}
\title{Get data for plotting from htmlReport object}
\arguments{
\item{options}{A list containing options for data retrieval.}
}
\value{
A list containing the retrieved data, attributes, samples, and
variables.
}
\description{
This method retrieves data suitable for plotting from an
\code{htmlReport} object based on specified options.
}
\details{
Get data for plotting from htmlReport object
}
\examples{
test_df <- data.frame(matrix(1, nrow = 5, ncol = 5))
test_df[1, ] <- letters[1:5]
test_df[, 1] <- letters[1:5]
plotter <- htmlReport$new(container = list(test_df = test_df))
plotter$get_data_for_plot(options = list(id = "test_df", transpose = FALSE, text = "dynamic"))
}
