% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_span-htmlReport-method}
\alias{get_span-htmlReport-method}
\title{get col and rowspan for a table element specified by indices.}
\arguments{
\item{colspan, rowspan}{table of colspans and rowspans.}

\item{row, col}{row and col indices of element whose spans to get.}
}
\value{
appended HTML code specifying element spans.

appended HTML code specifying element spans.
}
\description{
get_span

embed_html
}
\examples{
colspan_test <- data.frame(c(3, 2, 1), rep(1, 3), rep(1, 3))
rowspan_test <- data.frame(rep(1, 3), rep(1, 3), c(1, 2, 1))
plotter <- htmlReport$new()
plotter$get_span(colspan_test, rowspan_test, 1, 2)
}
