% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_heatmap.R
\name{gg_heatmap}
\alias{gg_heatmap}
\title{Function to draw heatmap, filling with provided parameter}
\usage{
gg_heatmap(
  input_table,
  x_axis = "x_axis",
  y_axis = "y_axis",
  fill = "Freq",
  text_plot = NULL,
  text_size = 2.5,
  text_colour = "black",
  xlabel = TRUE,
  ylabel = TRUE,
  x_angle = 25,
  transpose = FALSE,
  col = c("#0000D5", "#FFFFFF", "#D50000"),
  na_col = "grey50"
)
}
\arguments{
\item{input_table}{Table containing input input_table.}

\item{x_axis, y_axis}{Names of columns to use for x and y axes.}

\item{fill}{A string. Parameter with which the heatmap will be drawn.
Default: "Freq", for "frequency".}

\item{text_plot}{A string. Name of input_table column with text to
fill heatmap cells.}

\item{text_size}{A numeric. Font size of text_plot.}

\item{text_colour}{A string. Font colour of text_plot.}

\item{xlabel, ylabel}{Booleans.
\itemize{
\item \code{TRUE} (the default): Axis wil be labeled in plot.
\item \code{FALSE}:              Axes will NOT be labeled in plot.
}}

\item{x_angle}{A numeric. x axis label angle.}

\item{transpose}{A boolean.
\itemize{
\item \code{TRUE}:                Input will be transposed.
\item \code{FALSE} (the default): Input will NOT be transposed.
}}

\item{col}{A vector of colour code strings.}

\item{na_col}{A vector of colour code strings.}

\item{dendro}{Print dendrogram. Unused.}
}
\value{
A heatmap.
}
\description{
\code{gg_heatmap} takes a table (preferably in tidy format) and creates a heatmap
between two of its columns, filling blocks with provided parameter.
}
\examples{
none yet
}
