% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{\%+\%}
\alias{\%+\%}
\title{Custom addition operator for combining htmlReport objects}
\usage{
plotter \%+\% value
}
\arguments{
\item{plotter}{An object of class "htmlReport".}

\item{value}{An object of any type that can be coerced to character.}
}
\value{
An object of class "htmlReport" with an updated @all_report
which includes at the end the "value" string
}
\description{
This function defines a custom addition operator for combining two strings.
}
\examples{
\dontrun{
# Create two htmlReport objects
plotter <- new("htmlReport")
plotter <- plotter \%+\% "<h1>First Report</h1>"
} 
}
