% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{line-htmlReport-method}
\alias{line-htmlReport-method}
\title{Build CanvasXpress line plot from R data frame}
\arguments{
\item{options}{list with options.}
}
\value{
HTML code for CanvasXpress line plot of data.
}
\description{
Loads data frame and CanvasXpress options for line plot,
then calls canvasXpress_main to build it.
}
\details{
CanvasXpress line plot
}
\examples{
test_df <- data.frame(matrix(1, nrow = 5, ncol = 5))
test_df[1, ] <- letters[1:5]
test_df[, 1] <- letters[1:5]
plotter <- htmlReport$new(container = list(test_df = test_df))
plotter$line(list(id = "test_df", text = "dynamic"))
}
