% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_htmlreport.R
\name{main_htmlreportR}
\alias{main_htmlreportR}
\title{Build report by loading files from disk and rendering template.}
\usage{
main_htmlreportR(options)
}
\arguments{
\item{options}{list containing all necessary elements to build an
htmlReport instance.}
}
\value{
None
}
\description{
Main htmlreportR script mode function. Renders a template,
}
\details{
Main htmlreportR scripting function
}
\examples{
\dontrun{
options <- list(title = "main_htmlreportR example",
			   data_files = "path/to/file1", "path/to/fileN",
			   output_file = "path/to/output",
			   source_folder = "path/to/package/scripts",
			   compress_obj = TRUE,
			   css_files = "path/to/custom/css",
			   js_files = "path/to/custom/js",
			   cdn_css = "path/to/cdn_css",
			   cdn_js = "path/to/cdn_js",
			   menu = "menu")
main_htmlreportR (options)
}
}
