% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{make_head}
\alias{make_head}
\title{Make HTML Report Head}
\usage{
make_head(plotter)
}
\arguments{
\item{plotter}{An object of class \code{htmlReport}.}
}
\value{
An updated \code{htmlReport} object with the
title added to its head section.
}
\description{
This method generates the head section of an
HTML report by adding the title to an \code{htmlReport} object.
}
\examples{
\dontrun{
# Assuming plotter is an object of class htmlReport
plotter <- make_head(plotter)
}
}
