% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_html_list}
\alias{make_html_list}
\title{make_html_list}
\usage{
make_html_list(
  list_content,
  list_levels = NULL,
  list_types = NULL,
  default_type = "ul"
)
}
\arguments{
\item{list_content}{Vector or list of elements with which to build the list}

\item{list_levels}{Vector or list defining nesting levels. Default NULL}

\item{list_types}{Vector or list with types to assign to each element ("ul"
for unordered or "ol" for ordered).' If not supplied, default_type will
determine list type for each element. Default NULL}

\item{default_type}{List type to default if list_types is undefined, "ul"
for unordered or "ol" for ordered. Default "ul"}
}
\value{
Formatted html list ready to render.
}
\description{
Take an input vector or list and use it to build an html list of specified
type
}
\examples{
content <- c("One", "Two", "Three")
make_html_list(list_content = content)
}
