% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{merge_hashed_tables-htmlReport-method}
\alias{merge_hashed_tables-htmlReport-method}
\title{binds tables contained in hash_vars by rows}
\arguments{
\item{ids}{Vector of hash_vars IDs to bind.}

\item{join_method}{Method by which tables will be joined, default "rbind".
also admits "cbind", in which case from_id_name and alt_ids arguments will
be ignored.}

\item{add_colnames}{A boolean. Controls colnames handling.
\itemize{
\item \code{TRUE}: tables' colnames are treated as such.
\item \code{FALSE} (the default): tables' first row is moved to colnames.
}}

\item{from_id_name}{Name of column that contains original hash_vars IDs,
to trace their origin. If NULL, this column will not be added.}

\item{alt_ids}{New names for original IDs column, in case renaming them
is needed for clarity, for example. If NULL (the default) they will not be
renamed.}

\item{target_id}{Hash_vars ID where output will be saved. If NULL (the
default) it will be returned and NOT saved.}
}
\value{
Merged table
}
\description{
merge_hashed_tables
}
\examples{
container_test <- list(df_1 = data.frame(val = 1), df_2 = data.frame(val = 0))
plotter <- htmlReport$new(container = container_test, compress = FALSE)
plotter$merge_hashed_tables(ids = c("df_1", "df_2"))
}
