% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{merge_tables}
\alias{merge_tables}
\title{Merge tables from multiple hash_vars ids}
\arguments{
\item{options}{A list containing options for data retrieval.}
}
\value{
A merged data frame, minus specified fields and rows.
}
\description{
Non-exported method that will be called by extract_data if
it is provided with multiple ids.
}
\details{
Merge tables contained in hash_vars
}
