% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{parse_data_frame-htmlReport-method}
\alias{parse_data_frame-htmlReport-method}
\title{Print data frame in HTML format}
\arguments{
\item{data_frame}{Data frame to parse}

\item{options}{list with options}

\item{table_id}{An integer. Table id in report}

\item{row_names}{A boolean.
\itemize{
\item \code{TRUE} (the default): Parse data frame row names as a column of
the HTML table.
\item \code{FALSE} (the default): Do not parse data frame row names.
}}

\item{colspan, rowspan}{Mirrors of input data_frame specifying spans.}
}
\value{
A table in html format.
}
\description{
Parses a data frame included in an object of class "htmlReport"
and HTML table to include it in htmlreportR
}
\details{
Parse data frame to HTML
}
\examples{
test_df <- data.frame(matrix(1, nrow = 5, ncol = 5))
test_df[1, ] <- letters[1:5]
test_df[, 1] <- letters[1:5]
plotter <- htmlReport$new()
options <- list(border = 1, table_rownames = TRUE, rownames_col = "rownames")
plotter$parse_data_frame(test_df, options, "Sample")
}
