% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_in_div.R
\name{plot_in_div}
\alias{plot_in_div}
\title{Print R plot in a scrollable div}
\usage{
plot_in_div(
  g,
  fig_height = 7,
  fig_width = 12,
  cex = 1,
  max_size = 50,
  min_size = 10,
  counter = NULL
)
}
\arguments{
\item{cex, }{magnitude by which figure dimensions will be scaled}

\item{max_size, min_size}{plot area upper and lower limits in inches}

\item{counter, }{???}
}
\value{
html code containing a scrollable div with the printed plot.
}
\description{
\code{plot_in_div} allows a user to scroll through an html div to see a printed
plot without the need to scale it to div size.
}
\examples{
none yet
}
