% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_paired_mark}
\alias{replace_paired_mark}
\title{replace_paired_mark}
\usage{
replace_paired_mark(string, pattern, replace)
}
\arguments{
\item{string}{String to edit}

\item{pattern}{Pattern to recursively replace}

\item{replace}{Vector containing, in that order, expression that will replace
the first paired element and the second paired element}
}
\value{
Modified string.
}
\description{
Recursively replaces paired marks in a string.
}
\examples{
example_string <- "** This should be in bold **"
pattern <- "(\\\\*\\\\*+?)([- \\\\w]+)(\\\\*\\\\*+?)"
replace_paired_mark(example_string, pattern, c("<strong>", "</strong>"))
}
