% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{update_options-htmlReport-method}
\alias{update_options-htmlReport-method}
\title{Update reference options}
\arguments{
\item{ref}{list with original options}

\item{sub}{list with options to update}
}
\value{
A updated options list
}
\description{
This function updates the options of a option list based on other list
}
\details{
Update reference options
}
