% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{write_report-htmlReport-method}
\alias{write_report-htmlReport-method}
\title{Write HTML Report}
\arguments{
\item{output_path}{A character string specifying the output file path for the HTML report.}
}
\value{
Writes the HTML report to the specified output file path and removes temporary files.
}
\description{
This method writes the HTML report generated by an \code{htmlReport} object to a specified output file path.
}
\details{
Write HTML Report
}
\examples{
plotter <- htmlReport$new()
plotter$write_report(stdout())
}
