% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{write_report}
\alias{write_report}
\title{Write HTML Report}
\usage{
write_report(plotter, output_path)
}
\arguments{
\item{plotter}{An object of class \code{htmlReport}.}

\item{output_path}{A character string specifying}
}
\description{
Write HTML Report
}
