% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{barplot-htmlReport-method}
\alias{barplot-htmlReport-method}
\title{Build CanvasXpress barplot from R data frame}
\arguments{
\item{options}{list with options.}
}
\value{
HTML code for CanvasXpress bar plot of data.
}
\description{
Loads data frame and CanvasXpress options for barplot,
then calls canvasXpress_main to build it.
}
\details{
CanvasXpress barplot
}
