% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{build,htmlReport-method}
\alias{build,htmlReport-method}
\title{Build HTML report from template}
\usage{
\S4method{build}{htmlReport}(plotter, template)
}
\arguments{
\item{plotter}{An object of class "htmlReport".}

\item{template}{A character string specifying the file path of the template.}
}
\value{
An object of class "htmlReport" with HTML t built from the template.
}
\description{
This function builds an HTML report from a
given template file for an object of class "htmlReport".
}
\details{
This function reads the contents of the template file,
sets up Knitr options for rendering,
renders the template using Knitr, and constructs
the HTML report by combining the rendered
template with the existing report content in the "htmlReport" object.
}
\examples{
# Create an htmlReport object
\dontrun{
plotter <- new("htmlReport")

# Build the report from a template
plotter <- build(plotter, "template.html")
}
}
