% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{build-htmlReport-method}
\alias{build-htmlReport-method}
\title{Build HTML report from template}
\arguments{
\item{template}{A character string specifying the file path of the template.}
}
\value{
A new instance of the htmlreportR class
}
\description{
This function builds an HTML report from a given
template file for an object of class "htmlReport".
}
\details{
Build HTML report from template

This function reads the contents of the template file,
sets up Knitr options for rendering,
renders the template using Knitr, and constructs
the HTML report by combining the rendered
template with the existing report content in the "htmlReport" object.
}
