% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{build_body}
\alias{build_body}
\title{Build HTML Report Body}
\arguments{
\item{body_text}{A character string containing the body text
to be appended to the HTML report.}
}
\value{
An updated \code{htmlReport} object with the specified
body text appended to its body.
}
\description{
This method builds the body of an HTML report
by appending the specified body text to an \code{htmlReport} object.
}
\details{
Build HTML Report Body
}
