% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_to_rownames}
\alias{col_to_rownames}
\title{col_to_rownames}
\usage{
col_to_rownames(data_frame, col = 1)
}
\arguments{
\item{data_frame}{Data frame to manipulate.}

\item{col}{An integer. Col to set as new rowumns and remove from data frame.}
}
\description{
turns a column of the input data frame and turns it into its rownames, and
then removes the actual column.
}
\examples{
col_to_rownames(tail(mtcars))
}
