% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{concat}
\alias{concat}
\title{Custom addition operator for combining htmlReport objects}
\arguments{
\item{value}{An object of any type that can be coerced to character.}
}
\value{
An object of class "htmlReport" with an updated @all_report which includes at the end the "value" string.
}
\description{
This function defines a custom addition operator for combining two strings.
}
\details{
Custom addition operator for combining htmlReport objects
}
