% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{density-htmlReport-method}
\alias{density-htmlReport-method}
\title{Build CanvasXpress density plot from R data frame}
\arguments{
\item{options}{list with options.}
}
\value{
HTML code for CanvasXpress density plot of data.
}
\description{
Loads data frame and CanvasXpress options for density plot,
then calls canvasXpress_main to build it.
}
\details{
CanvasXpress density plot
}
