% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_id_column}
\alias{.add_id_column}
\title{.add_id_column}
\usage{
.add_id_column(id, hash_vars, from_id_name)
}
\arguments{
\item{id}{hash_vars ID of table to process}

\item{hash_vars}{plotter hash_vars slot}

\item{from_id_name}{Name of column that contains original hash_vars IDs,
to trace their origin. If NULL, this column will not be added.}
}
\description{
takes hash_vars, looks for a specific table, and adds its ID in a new
column.
}
