% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{extract_data}
\alias{extract_data}
\title{Extract Data from htmlReport Object dataframe}
\usage{
extract_data(plotter, data_frame, options)
}
\arguments{
\item{plotter}{An object of class \code{htmlReport}.}

\item{data_frame}{A data frame containing the data to be extracted.}

\item{options}{A list containing options for data extraction.
#'}
}
\value{
A list containing the extracted data and additional information.
}
\description{
This method extracts data from an \code{htmlReport}
object based on specified options.
}
\examples{
\dontrun{
plotter <- new("htmlReport")
data_frame <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6), C = c(7, 8, 9))
options <- list(smp_attr = NULL,
                var_attr = NULL,
                fields = NULL)
extracted_data <- extract_data(plotter, data_frame, options)
}

}
