% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{extract_data}
\alias{extract_data}
\title{Retrieve Data from htmlReport Object}
\arguments{
\item{options}{A list containing options for data retrieval.}
}
\value{
A list containing the retrieved data and additional information.
}
\description{
This method retrieves data from an \code{htmlReport} object based on specified options.
}
\details{
Retrieve Data from htmlReport Object
}
