% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_data_for_plot}
\alias{get_data_for_plot}
\title{Get Data for Plotting from htmlReport Object}
\arguments{
\item{options}{A list containing options for data retrieval.}
}
\value{
A list containing the retrieved data, attributes, samples, and variables.
}
\description{
This method retrieves data suitable for plotting from an \code{htmlReport} object based on specified options.
}
\details{
Get Data for Plotting from htmlReport Object
}
\examples{
\dontrun{
# Assuming plotter is an object of class htmlReport
data_frame <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6), C = c(7, 8, 9))
options <- list(id = "data_id", transpose = FALSE)
data <- plotter$get_data_for_plot(options)
}
}
