% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_data_for_plot}
\alias{get_data_for_plot}
\title{Get Data for Plotting from htmlReport Object}
\arguments{
\item{options}{A list containing options for data retrieval.}
}
\value{
A list containing the retrieved data, attributes, samples, and variables.
}
\description{
This method retrieves data suitable for plotting from an \code{htmlReport} object based on specified options.
}
\details{
Get Data for Plotting from htmlReport Object
}
