% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_plot}
\alias{get_plot}
\title{Get Plot from htmlReport Object}
\arguments{
\item{plot_obj}{A plot object like ggplot.}

\item{width}{plot width}

\item{height}{plot height}

\item{size_unit}{units in plot size}

\item{img_properties}{string including html properties of img}

\item{resizable}{logical indicating if plot must be resizable}
}
\value{
Displays the plot within the HTML report.
}
\description{
This method generates and retrieves a plot from an \code{htmlReport} object. This code writes the plot to a temporal png, then it loads the png in base64 encoding and then displays the plot within the HTML report.
}
\details{
Get Plot from htmlReport Object
}
