% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Get js Plot from canvasxpress_obj Object}
\arguments{
\item{object_id}{string indicating object id}

\item{cvXpress}{vanvasXpress_obj object}
}
\value{
Displays the js code for plot.
}
\description{
get_plot_data
}
