% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{get_span-htmlReport-method}
\alias{get_span-htmlReport-method}
\title{get col and rowspan for a table element specified by indices.}
\arguments{
\item{colspan, rowspan}{table of colspans and rowspans.}

\item{row, col}{row and col indices of element whose spans to get.}
}
\value{
appended HTML code specifying element spans.
}
\description{
embed_html
}
