% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{parse_data_frame-htmlReport-method}
\alias{parse_data_frame-htmlReport-method}
\title{Print data frame in HTML format}
\arguments{
\item{df}{Data frame to parse}

\item{id}{An integer. Table id in report}

\item{border}{An integer. Border thickness}

\item{row_names}{A boolean.
\itemize{
\item \code{TRUE} (the default): Parse data frame row names as a column of
the HTML table.
\item \code{FALSE} (the default): Do not parse data frame row names.
}}
}
\value{
A table in html format.
}
\description{
Parses a data frame included in an object of class "htmlReport"
and HTML table to include it in htmlreportR
}
\details{
Parse data frame to HTML
}
