% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_paths}
\alias{parse_paths}
\title{parse_paths}
\usage{
parse_paths(string)
}
\arguments{
\item{string}{String containing paths to expand.}
}
\description{
takes a vector of comma-separated paths and parses it so files can be loaded.
If any of the paths contain wildcards, they will be expanded.
}
