% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{plot_grid-htmlReport-method}
\alias{plot_grid-htmlReport-method}
\title{represents a list of plots in a grid}
\arguments{
\item{plots}{List of plots to grid}

\item{width}{Grid width}

\item{height}{Grid height}
}
\value{
Grid including every plot
}
\description{
plot_grid
}
