% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{prettify_div-htmlReport-method}
\alias{prettify_div-htmlReport-method}
\title{Encapsulates input code in div of specified characteristics}
\arguments{
\item{data}{string indicating object id}
}
\value{
encoded and compressed json
}
\description{
pretiffy_div
}
