% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_to_header}
\alias{row_to_header}
\title{row_to_header}
\usage{
row_to_header(data_frame, row = 1)
}
\arguments{
\item{data_frame}{Data frame to manipulate.}

\item{row}{An integer. Row to set as new columns and remove from data frame.}
}
\description{
turns a row of the input data frame and turns it into its colnames, and then
removes the actual row.
}
\examples{
row_to_header(head(mtcars))
}
