% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_to_header.R
\name{row_to_header}
\alias{row_to_header}
\title{Select row from data frame and make it its header}
\usage{
row_to_header(df, row = 1, update_rows = TRUE)
}
\arguments{
\item{df}{A data frame of N rows.}

\item{row}{An integer. Index of row to turn into header (default 1).}

\item{update_rows}{A boolean.
\itemize{
\item \code{TRUE} (the default): data frame row names will be updated (first column
will be 1, second will be 2). If input df had no
row names, they are coerced to numerical values.
This option ensures they will reflect their new
positions.
\item \code{FALSE}:              data frame row names will be left untouched.
useful if they had previous values.
}}
}
\value{
A data frame of N-1 rows, where the selected row is now the header.
}
\description{
\code{parse_column_names} takes an input data frame and renames set columns in
order to simplify manipulation by other DRAGGER functions.
}
\examples{
df <- data.frame(c("Animal", "Fish"), c("Item", "Table"))
print(df)
header_df <- row_to_header(df, 1)
print(header_df)
}
