% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{scatter2D-htmlReport-method}
\alias{scatter2D-htmlReport-method}
\title{Build CanvasXpress scatter2D plot from R data frame}
\arguments{
\item{options}{list with options.}
}
\value{
HTML code for CanvasXpress scatter2D plot of data.
}
\description{
Loads data frame and CanvasXpress options for scatter2D plot,
then calls canvasXpress_main to build it.
}
\details{
CanvasXpress scatter2D plot
}
