% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{scatter3D-htmlReport-method}
\alias{scatter3D-htmlReport-method}
\title{Build CanvasXpress scatter3D plot from R data frame}
\arguments{
\item{options}{list with options.}
}
\value{
HTML code for CanvasXpress scatter3D plot of data.
}
\description{
Loads data frame and CanvasXpress options for scatter3D plot,
then calls canvasXpress_main to build it.
}
\details{
CanvasXpress scatter3D plot
}
