% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{static_ggplot_main,htmlReport-method}
\alias{static_ggplot_main,htmlReport-method}
\title{Generate static ggplot for HTML report}
\usage{
\S4method{static_ggplot_main}{htmlReport}(
  plotter,
  id,
  header = FALSE,
  row_names = FALSE,
  transpose = FALSE,
  smp_attr = NULL,
  var_attr = NULL,
  fields = NULL,
  func = NULL,
  plotting_function = NULL
)
}
\arguments{
\item{plotter}{An object of class "htmlReport".}

\item{id}{A character string specifying the
identifier for the element of hash_vars that is taken.}

\item{header}{Logical, indicating whether the dataset has a header row.}

\item{row_names}{Logical, indicating whether to include row names.}

\item{transpose}{Logical, indicating whether to transpose the dataset.}

\item{smp_attr}{A list of attributes for samples.}

\item{var_attr}{A list of attributes for variables.}

\item{fields}{A character vector specifying  fields to include in the ggplot.}

\item{func}{A function to preprocess data before plotting.}

\item{plotting_function}{A function used for generating the ggplot.}
}
\value{
An object of class "htmlReport" with the static ggplot included.
}
\description{
This function generates a static ggplot for inclusion in an
HTML report for an object of class "htmlReport".
}
\details{
This function generates a static ggplot based on the
provided data and plot specifications. It first
defines a wrapper function for ggplot generation, then
calls the \code{static_plot_main} function
to generate the plot and include it in the HTML report object.
}
