% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{static_ggplot_main-htmlReport-method}
\alias{static_ggplot_main-htmlReport-method}
\title{Generate static ggplot for HTML report}
\arguments{
\item{id}{A character string specifying the identifier
for the element of hash_vars that is taken.}

\item{header}{Logical, indicating whether the dataset has a header row.}

\item{row_names}{Logical, indicating whether to include row names.}

\item{transpose}{Logical, indicating whether to transpose the dataset.}

\item{smp_attr}{A list of attributes for samples.}

\item{var_attr}{A list of attributes for variables.}

\item{fields}{A character vector specifying fields to include in the ggplot.}

\item{func}{A function to preprocess data before plotting.}

\item{plotting_function}{A function used for generating the ggplot.}

\item{text}{Logical, indicating whether to convert table to text or a vector indicating the numeric fields.}

\item{width}{plot width}

\item{height}{plot height}

\item{size_unit}{units in plot size}

\item{img_properties}{string including html properties of img}

\item{resizable}{logical indicating if plot must be resizable}
}
\description{
This function generates a static ggplot for
inclusion in an HTML report for an object of class "htmlReport".
}
\details{
Generate static ggplot for HTML report

This function generates a static ggplot based on the provided data
and plot specifications. It first defines a wrapper function for ggplot
generation, then calls the \code{static_plot_main} function to generate the
plot and include it in the HTML report object.
}
