% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{static_plot_main-htmlReport-method}
\alias{static_plot_main-htmlReport-method}
\title{Generate static plot for HTML report}
\arguments{
\item{id}{A character string specifying the identifier for the plot included
in hash_vars.}

\item{header}{Logical, indicating whether the dataset has a header row.}

\item{row_names}{Logical, indicating whether to include row names.}

\item{transpose}{Logical, indicating whether to transpose the dataset.}

\item{smp_attr}{A list of attributes for samples.}

\item{var_attr}{A list of attributes for variables.}

\item{fields}{A character vector specifying the fields to include in the
plot.}

\item{func}{A function to preprocess data before plotting.}

\item{plotting_function}{A function used for generating the plot.}

\item{plotting_args}{Named list of arguments to pass to plotting function in
addition to input data.}

\item{text}{Logical, indicating whether to convert table to text or a vector
indicating the numeric fields.}

\item{custom_format}{Logical, indicating if id correspond to a table or a
custom object}

\item{width}{plot width}

\item{height}{plot height}

\item{size_unit}{units in plot size}

\item{img_properties}{string including html properties of img}

\item{resizable}{logical indicating if plot must be resizable}
}
\description{
This function generates a static plot for inclusion in an HTML
report for an object of class "htmlReport".
}
\details{
Generate static plot for HTML report

This function generates a static plot based on the provided data and plot
specifications.
It first retrieves the data frame for plotting using the provided options,
preprocesses the
data if a preprocessing function is specified, generates the plot using the
provided plotting function, and then adds the plot to the HTML report object.
}
\examples{
test_func <- function(input_obj, columns) {
	p <- input_obj[, columns, drop = FALSE]
	print(p)
	return(p)
}
test_df <- data.frame(A = 1:5, B= 6:10, attr = "potato")
plotter <- htmlReport$new(container = list(df1 = test_df))
output <- plotter$static_plot_main(id = "df1",
	text = "dynamic", plotting_function = test_func, plot_type = "plot",
	plotting_args = list(columns = 1))
}
