% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{static_plot_main,htmlReport-method}
\alias{static_plot_main,htmlReport-method}
\title{Generate static plot for HTML report}
\usage{
\S4method{static_plot_main}{htmlReport}(
  plotter,
  id,
  header = FALSE,
  row_names = FALSE,
  transpose = FALSE,
  smp_attr = NULL,
  var_attr = NULL,
  fields = NULL,
  func = NULL,
  plotting_function = NULL
)
}
\arguments{
\item{plotter}{An object of class "htmlReport".}

\item{id}{A character string specifying the
identifier for the plot included in hash_vars.}

\item{header}{Logical, indicating whether the dataset has a header row.}

\item{row_names}{Logical, indicating whether to include row names.}

\item{transpose}{Logical, indicating whether to transpose the dataset.}

\item{smp_attr}{A list of attributes for samples.}

\item{var_attr}{A list of attributes for variables.}

\item{fields}{A character vector specifying the fields to include in the plot}

\item{func}{A function to preprocess data before plotting.}

\item{plotting_function}{A function used for generating the plot.}
}
\value{
An object of class "htmlReport" with the static plot included.
}
\description{
This function generates a static plot for inclusion in an
HTML report for an object of class "htmlReport".
}
\details{
This function generates a static plot based on
the provided data and plot specifications. It first
retrieves the data frame for plotting using the provided options,
preprocesses the data if a preprocessing
function is specified, generates the plot using the provided
plotting function, and then adds the plot
to the HTML report object.
}
