% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{table}
\alias{table}
\title{Generate table from plotter object with specified options}
\arguments{
\item{options}{list with options}
}
\value{
A string encoding a html table.
}
\description{
This method takes a list of options to build a table from a
htmlReport object.
}
\details{
Table method for htmlReport class
}
\examples{
test_df <- data.frame(matrix(1, nrow = 5, ncol = 5))
test_df[1, ] <- letters[1:5]
test_df[, 1] <- letters[1:5]
plotter <- htmlReport$new(container = list(test_df = test_df))
plotter$table(list(id = "test_df", header = TRUE,
									row_names = TRUE, smp_attr = 2,
									var_attr = 2))
}
