% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreport.R
\name{table}
\alias{table}
\title{Generate table from plotter object with specified options}
\arguments{
\item{options}{list with options}
}
\value{
A string encoding a html table.
}
\description{
This method takes a list of options to build a table from a
htmlReport object.
}
\details{
Table method for htmlReport class
}
